Added support for IP lookups
You can now lookup the hostname for a given IP address. Supports both IPv4 and IPv6.

This will be a separate lookup to the domain one, and to assist with this and help you differentiate between the two lookups, the supplied (Optional) HTACCESS file has been updated with new RewriteRules to reflect this.

Perform a GET request to the API lookup.php file with the param ip = [the ip to lookup], and you will get a JSON response. The IP address you looked up, and the corresponding hostname.

Updated URLs if using HTACCESS are...

Domain Lookups: /lookup/domain/one.one.one.one
IP Lookups: /lookup/ip/1.1.1.1

The above uses Cloudflare’s one.one.one.one and 1.1.1.1 as examples. Replace these with the appropriate values for your lookup request.